% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjs_params.R
\name{cjs_params}
\alias{cjs_params}
\title{Compute survival and capture probabilities from a fitted multilevel CJS model}
\usage{
cjs_params(object, newX, newgroups, N_mcmc = NULL)
}
\arguments{
\item{object}{Fitted model object of class \code{stanfit} representing a
multilevel CJS model with \code{pars} \code{beta}, \code{epsilon}, \code{b}, and \code{e} (see XXX
for definitions).}

\item{newX}{New \verb{M x K} model matrix, as returned by \code{model.matrix},
containing exactly the same predictors, in the same order, as were passed
as data variable \code{X} to \code{object}. Use the same \code{X} to calculate \code{phi} and
\code{p} for the fitted data.}

\item{newgroups}{List with elements named \code{phi} and \code{p}, containing the
matrices (\verb{M x (T-1)} and \verb{M x T}, respectively) of integer-valued group
indices corresponding to \code{newX}. The unique indices for each occasion
(column) must be a subset of those used for the corresponding occasion of
\code{group_phi} or \code{group_p} passed to \code{object}. If there is only one group for
a given parameter at a given occasion, group-varying random effects will
not be used. Use the original \code{group_phi} and \code{group_p} to calculate \code{phi}
and \code{p} for the fitted data.}

\item{N_mcmc}{Number of posterior draws to return. If \code{NULL} (the default),
all samples in \code{object} are used.}
}
\value{
A list whose elements \code{phi} and \code{p} are arrays (\verb{N_mcmc x M x (T-1)}
and \verb{N_mcmc x M x T}, respectively) containing the posterior draws.
}
\description{
Compute survival and capture probabilities from a fitted multilevel CJS model
}
