% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJStan-package.R
\docType{package}
\name{CJStan-package}
\alias{CJStan-package}
\alias{CJStan}
\title{The 'CJStan' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.26.13. https://mc-stan.org
}
